refactor(adapters): import kernel through barrel
