+++ packages/adapters/local/src/index.ts
@@ -2,13 +2,13 @@ import { execFileSync } from "node:child_process";
 import { existsSync } from "node:fs";
 import path from "node:path";
 import { Effect } from "effect";
-import type { EngineError, WriteError } from "../../../kernel/src/domain/index.ts";
-import { explainStatusTransition, isTerminalStatus } from "../../../kernel/src/domain/index.ts";
-import { evaluateEntityDisposition } from "../../../kernel/src/entity/disposition.ts";
-import { stablePayloadHash } from "../../../kernel/src/integrity/stable-hash.ts";
-import type { HarnessLayoutInput } from "../../../kernel/src/layout/index.ts";
-import { createHarnessRuntimeContext, harnessRuntimeRoot, taskPackagePath } from "../../../kernel/src/layout/index.ts";
-import type { WriteCoordinator } from "../../../kernel/src/ports/index.ts";
+import type { EngineError, WriteError } from "../../../kernel/src/index.ts";
+import { explainStatusTransition, isTerminalStatus } from "../../../kernel/src/index.ts";
+import { evaluateEntityDisposition } from "../../../kernel/src/index.ts";
+import { stablePayloadHash } from "../../../kernel/src/index.ts";
+import type { HarnessLayoutInput } from "../../../kernel/src/index.ts";
+import { createHarnessRuntimeContext, harnessRuntimeRoot, taskPackagePath } from "../../../kernel/src/index.ts";
+import type { WriteCoordinator } from "../../../kernel/src/index.ts";
 import { makeJournaledWriteCoordinator, runLedgerMaterializer } from "../../../kernel/src/store/index.ts";
 import { resolveTaskCreatedBy } from "./created-by.ts";
 import { renderSupersedesRelation } from "./task-relations.ts";
