+++ packages/adapters/local/test/task-writes.test.ts
@@ -1,8 +1,8 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 import { Effect } from "effect";
-import { taskEntityId } from "../../../kernel/src/domain/index.ts";
-import type { WriteCoordinator, WriteOp } from "../../../kernel/src/ports/index.ts";
+import { taskEntityId } from "../../../kernel/src/index.ts";
+import type { WriteCoordinator, WriteOp } from "../../../kernel/src/index.ts";
 import { writeSupersedeTaskDocuments } from "../src/task-writes.ts";
 
 test("supersede document writes use the explicit operation task id", () => {
